Telecom: Ues proper sub IDs
It is assumed that the ID of every PhoneAccountHandle is the string
equivalent of an int, however, the ID can also includes hex chars or,
in case of the emergency account handle, it's just "E". This causes
issues when we try to get the sub ID using Integer.parseInt(),
so don't do it.

Change-Id: If82c8146d01dba03cc3323aef9a4df84dc1f4bfd
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>